Add explicit cast in half_btf()
To silence asan failures in fuzzing tests.

BUG=:68825590,68825594,68825599

+++ av1/common/av1_txfm.h
@@ -103,7 +103,7 @@ static INLINE void round_shift_array(int32_t *arr, int size, int bit) {
 
 static INLINE int32_t half_btf(int32_t w0, int32_t in0, int32_t w1, int32_t in1,
                                int bit) {
-  int32_t result_32 = w0 * in0 + w1 * in1;
+  int32_t result_32 = (int32_t)((int64_t)w0 * in0 + w1 * in1);
 #if CONFIG_COEFFICIENT_RANGE_CHECKING
   int64_t result_64 = (int64_t)w0 * (int64_t)in0 + (int64_t)w1 * (int64_t)in1;
   if (result_64 < INT32_MIN || result_64 > INT32_MAX) {
